products post/put/delete jwt check

+++ src/app/api/products/[id]/route.ts
@@ -1,4 +1,5 @@
 import { prisma } from "@/db";
+import { isAuthorized } from "@/utils/auth";
 import { Prisma } from "@prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -22,6 +23,15 @@ export async function DELETE(req: NextRequest, { params }: Slug) {
   if (!params.id)
     return NextResponse.json({ error: "not found" }, { status: 404 });
 
+  // jwt check start
+  const authorization = req.headers.get("authorization");
+  console.log(authorization);
+
+  const validUser = isAuthorized(authorization);
+  if (!validUser)
+    return NextResponse.json({ error: "unauthorize" }, { status: 404 });
+  // jwt check end
+
   const deletedProduct = await prisma.product.delete({
     where: { id: params.id },
   });
@@ -37,6 +47,15 @@ export async function PUT(req: NextRequest, { params }: Slug) {
 
   const data = (await req.json()) as Prisma.ProductUpdateInput;
   console.log(data);
+
+  // jwt check start
+  const authorization = req.headers.get("authorization");
+  console.log(authorization);
+
+  const validUser = isAuthorized(authorization);
+  if (!validUser)
+    return NextResponse.json({ error: "unauthorize" }, { status: 404 });
+  // jwt check end
 
   const editedProduct = await prisma.product.update({
     where: { id: params.id },
